yeet

+++ CISC_327/Profile_Page/Payment_Method/Delete/Delete_Payment_Method.py
@@ -6,10 +6,10 @@ delete_payment_blueprint = Blueprint('delete_payment',__name__)
 def delete_payment():
     payment_id = request.form.get('payment_id')
     payment_method = Payment_Method.query.get(payment_id)
-    if payment_method.user_id == current_user.id:
+    if payment_method.user_id == current_user.id and len(current_user.payment_methods)>1:
         db.session.delete(payment_method)
         db.session.commit()
-        flash('Address deleted successfully!', 'success')
+        flash('Payment Method deleted successfully!', 'success')
     else:
         flash('Unauthorized', 'danger')
     return redirect(url_for('profile.profile'))
